Update
Add SSL option, switch to using repository for php7.0-memcached, Add
checks

+++ php7.sh
@@ -6,6 +6,7 @@ fi
 STARTNGINX='service nginx restart'
 STARTXBT='./xbt_tracker'
 STARTMEMCACHED='service memcached restart'
+STARTPHPFPM='service php7.0-fpm restart'
 user='u232'
 db='u232'
 dbhost='localhost'
@@ -14,7 +15,6 @@ announcebase='http:\/\/'
 httpsannouncebase='https:\/\/'
 announce2='\/announce.php'
 xbt='xbt'
-codename=$(lsb_release -a | grep Codename | awk '{ printf $2 }')
 function randomString {
         local myStrLength=16;
         local mySeedNumber=$$`date +%N`;
@@ -33,10 +33,14 @@ echo -n "Enter the site's name: "
 read name
 echo -n "Enter the site's email: "
 read email
+echo -n "Do you want to enable SSL (y/n): "
+read ssl
 announce=$announcebase$baseurl$announce2
 httpsannounce=$httpsannouncebase$baseurl$announce2
 apt-get -y update
 apt-get -y upgrade
+apt-get -y install lsb-release
+codename=$(lsb_release -a | grep Codename | awk '{ printf $2 }')
 
 case $codename in
 	"jessie")
@@ -47,15 +51,14 @@ case $codename in
         exit 1
         ;;
 esac
-if [[ $xbtyn = 'n' ]]; then
-    exit 1
-fi
 if [[ $xbt = 'xbt' ]]; then
     echo -n "Do you want to run XBT tracker or php? (xbt/php)"
     read xbt
 fi
 case $xbt in
     'xbt')
+		extras='libmariadbclient-dev libpcre3 libpcre3-dev cmake g++ libboost-date-time-dev libboost-dev libboost-filesystem-dev libboost-program-options-dev libboost-regex-dev libboost-serialization-dev make subversion zlib1g-dev'
+		announce=$announcebase$baseurl
         ;;
     'php')
         ;;
@@ -65,24 +68,18 @@ case $xbt in
         exit 1
         ;;
 esac
+
 apt-get -y install software-properties-common
 apt-key adv --recv-keys --keyserver hkp://keyserver.ubuntu.com:80 0xcbcb082a1bb943db
 wget https://www.dotdeb.org/dotdeb.gpg
 apt-key add dotdeb.gpg
 rm dotdeb.gpg
-if [[ $codename != "squeeze" ]]; then
-	add-apt-repository "deb [arch=amd64,i386] http://nyc2.mirrors.digitalocean.com/mariadb/repo/10.1/debian jessie main"
-    add-apt-repository "deb http://packages.dotdeb.org jessie all"
-fi
+add-apt-repository "deb [arch=amd64,i386] http://nyc2.mirrors.digitalocean.com/mariadb/repo/10.1/debian jessie main"
+add-apt-repository "deb http://packages.dotdeb.org jessie all"
 
 apt-get -y update
-apt-get -y install mariadb-server memcached unzip libssl-dev php7.0 php7.0-mysql php7.0-json locate php7.0-fpm php7.0-dev git pkg-config build-essential nginx libmemcached-dev
+apt-get -y install mariadb-server memcached unzip libssl-dev php7.0 php7.0-mysql php7.0-json locate php7.0-fpm nginx php7.0-memcached $extras
 
-
-if [[ $xbt = 'xbt' ]]; then
-    announce=$announcebase$baseurl
-    apt-get -y install libmariadbclient-dev libpcre3 libpcre3-dev cmake g++ libboost-date-time-dev libboost-dev libboost-filesystem-dev libboost-program-options-dev libboost-regex-dev libboost-serialization-dev make subversion zlib1g-dev 
-fi
 updatedb
 mysql_secure_installation
 cd /etc/nginx/sites-enabled
@@ -92,6 +89,7 @@ sed -i "s/group = www-data/group = www-data/" /etc/php/7.0/fpm/pool.d/www.conf
 sed -i "s/;listen\.owner.*/listen.owner = www-data/" /etc/php/7.0/fpm/pool.d/www.conf
 sed -i "s/;listen\.group.*/listen.group = www-data/" /etc/php/7.0/fpm/pool.d/www.conf
 sed -i "s/;listen\.mode.*/listen.mode = 0660/" /etc/php/7.0/fpm/pool.d/www.conf # This passage in not required normally
+echo "memcached.serializer = 'php'" >> /etc/php/7.0/fpm/php.ini
 rm default*
 cd ../sites-available
 rm default*
@@ -115,7 +113,7 @@ echo "server {
 
         fastcgi_index index.php;
         include fastcgi_params;
-        fastcgi_param SCRIPT_FILENAME $document_root$fastcgi_script_name;
+        fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
     }
 }" > /etc/nginx/sites-available/default
 ln -s /etc/nginx/sites-available/default /etc/nginx/sites-enabled
@@ -158,8 +156,8 @@ echo > /var/www/include/settings/settings.txt
 chmod 777 /var/www/include/settings/settings.txt
 chmod 777 /var/www/sqlerr_logs/
 chmod 777 /var/www/torrents
-rm /var/www/include/classes/class_cache.php
-wget https://gitlab.open-scene.net/whocares/u232-v4-xbt-mariadb-php5/raw/master/class_cache.php -O /var/www/include/classes/class_cache.php
+rm /var/www/include/class/class_cache.php
+wget https://gitlab.open-scene.net/whocares/u232-v4-xbt-mariadb-php5/raw/master/class_cache.php -O /var/www/include/class/class_cache.php
 configfile='/var/www/install/extra/config.'$xbt'sample.php'
 sed 's/#mysql_user/'$user'/' $configfile > /var/www/include/config.php
 sed -i 's/#mysql_pass/'$pass'/' /var/www/include/config.php
@@ -187,27 +185,62 @@ chown -R www-data:www-data /var/www
 chown -R www-data:www-data /var/bucket
 
 cd ~
-git clone https://github.com/php-memcached-dev/php-memcached.git
-cd php-memcached
-git checkout php7
-phpize
-./configure --disable-memcached-sasl
-make
-make install
 if [ ! -f /etc/php/mods-available/memcached.ini ]; then
 echo "; configuration for php memcached module
 ; priority=20
 extension=memcached.so" > /etc/php/mods-available/memcached.ini
 fi
+if [ ! -f /etc/php/7.0/fpm/conf.d/20-memcached.ini ]; then
 ln -s /etc/php/mods-available/memcached.ini /etc/php/7.0/fpm/conf.d/20-memcached.ini
+fi
+if [ ! -f /etc/php/7.0/cli/conf.d/20-memcached.ini ]; then
 ln -s /etc/php/mods-available/memcached.ini /etc/php/7.0/cli/conf.d/20-memcached.ini
-service php7.0-fpm restart
-$STARTNGINX
-
+fi
+$STARTPHPFPM
 cd ~
 
+if [[ $ssl = 'y' ]]; then
+	apt-get install -y openssl
+	mkdir -p /etc/nginx/ssl/cert
+	mkdir -p /etc/nginx/ssl/private
+	openssl genrsa -des3 -out $baseurl.key 2048
+	openssl req -new -key $baseurl.key -out $baseurl.csr
+	cp $baseurl.key $baseurl.key.org
+	openssl rsa -in $baseurl.key.org -out $baseurl.key
+	rm $baseurl.key.org
+	openssl x509 -req -days 365 -in $baseurl.csr -signkey $baseurl.key -out $baseurl.crt
+	cp $baseurl.crt /etc/nginx/ssl/cert/
+	cp $baseurl.key /etc/nginx/ssl/private/
+	echo "server {
+    listen   443;
+    ssl on;
+    ssl_certificate /etc/nginx/ssl/cert/$baseurl.crt;
+    ssl_certificate_key /etc/nginx/ssl/private/$baseurl.key;
+    server_name $baseurl;
+    root /var/www;
+    index index.html index.htm index.php;
 
+    server_name $baseurl;
 
+    location / {
+        try_files \$uri \$uri/ /index.php\$is_args\$args;
+    }
+
+    location ~ \.php\$ {
+        fastcgi_split_path_info ^(.+\.php)(/.+)\$;
+
+        # With php7.0-fpm:
+        fastcgi_pass unix:/run/php/php7.0-fpm.sock;
+
+        fastcgi_index index.php;
+        include fastcgi_params;
+        fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
+    }
+}" > /etc/nginx/sites-available/$baseurl-ssl
+ln -s /etc/nginx/sites-available/$baseurl-ssl /etc/nginx/sites-enabled
+fi
+
+$STARTNGINX
 
 if [[ $xbt = 'xbt' ]]; then
     svn co -r 2466 http://xbt.googlecode.com/svn/trunk/xbt/misc xbt/misc
@@ -266,6 +299,62 @@ else
     then 
 
     $STARTMEMCACHED
+
+        if ps ax | grep -v grep | grep $SERVICE >/dev/null
+
+    then
+
+        echo "$SERVICE service is now restarted, everything is fine"
+
+        fi
+
+    fi
+        
+fi
+######CHECK nginx######
+SERVICE='nginx'
+
+ if  ps ax | grep -v grep | grep $SERVICE > /dev/null
+then
+    echo "$SERVICE service running, everything is fine"
+else
+    echo "$SERVICE is not running, restarting $SERVICE" 
+
+     chkmem=`ps ax | grep -v grep | grep -c memcached`
+
+     if [ $chkmem <= 0 ]
+
+    then 
+
+    $STARTNGINX
+
+        if ps ax | grep -v grep | grep $SERVICE >/dev/null
+
+    then
+
+        echo "$SERVICE service is now restarted, everything is fine"
+
+        fi
+
+    fi
+        
+fi
+######CHECK php-fpm######
+SERVICE='php-fpm'
+
+ if  ps ax | grep -v grep | grep $SERVICE > /dev/null
+then
+    echo "$SERVICE service running, everything is fine"
+else
+    echo "$SERVICE is not running, restarting $SERVICE" 
+
+     chkmem=`ps ax | grep -v grep | grep -c memcached`
+
+     if [ $chkmem <= 0 ]
+
+    then 
+
+    $STARTPHPFPM
 
         if ps ax | grep -v grep | grep $SERVICE >/dev/null
